+++ src/codegen/compile_classical.ml
@@ -8055,13 +8055,24 @@ module MakeSerialization (Strm : Stream) = struct
       | Prim Blob ->
         with_blob_typ env (read_blob ())
       | Prim Principal ->
-        with_prim_typ t
-        begin
+        (* rule: `service <actortype> <: principal`, so also accept a service reference *)
+        let read_principal_data () =
           read_byte_tagged
             [ E.trap_with env "IDL error: unexpected principal reference"
             ; read_principal Tagged.P ()
             ]
-        end
+        in
+        check_prim_typ t ^^
+        G.if1 I32Type
+          (read_principal_data ())
+          begin
+            check_composite_typ get_idltyp idl_service ^^
+            G.if1 I32Type
+              (read_principal_data ())
+              ( skip get_idltyp ^^
+                coercion_failed ("IDL error: unexpected IDL type when parsing " ^ string_of_typ t)
+              )
+          end
       | Prim Text ->
         with_prim_typ t (read_text ())
       | Tup [] -> (* e(()) = null *)